+++ Smart-home-box-by-Hamza/front-back/SmartHomeApi/SmartHomeApi/Controllers/SkeniraniPokretiController.cs
@@ -7,6 +7,7 @@ using SmartHomeApi.Helpers;
 namespace SmartHomeApi.Controllers
 {
     [ApiController]
+    [Auth]
     [Route("[controller]/[action]")]
     public class SkeniraniPokretiController
     {
@@ -14,18 +15,22 @@ namespace SmartHomeApi.Controllers
         private readonly WhatsAppSender sender;
         private readonly SmsSender smsSender;
         private readonly EmailSender emailSender;
+        private readonly AuthService auth;
         public SkeniraniPokretiController(MyDBContext _db, WhatsAppSender _sender, SmsSender _smsSender,
-            EmailSender _emailSender)
+            EmailSender _emailSender, AuthService _auth)
         {
             db = _db;
             sender = _sender;
             smsSender = _smsSender;
             emailSender = _emailSender;
+            auth = _auth;
         }
         [HttpPost]
         public async Task<bool> addPokret()
         {
-            var pokret = new SkeniraniPokret { DatumVrijeme = DateTime.Now};
+            var prijavaInfo = await auth.getInfo();
+            var pokret = new SkeniraniPokret { DatumVrijeme = DateTime.Now
+                , KorisnikId = prijavaInfo.Prijava.KorisnikId};
             db.SkeniraniPokreti.Add(pokret);
             db.SaveChanges();
             var poruka = "Smart home box by Hamza: Skeniran pokret u prostoru. Molimo posjetite aplikaciju za vise detalja.";
@@ -37,7 +42,9 @@ namespace SmartHomeApi.Controllers
         [HttpGet]
         public async Task<PokretiResponse> getPokretiAll ()
         {
-            List<PokretResponse> list = await db.SkeniraniPokreti.OrderByDescending(sp=> sp.DatumVrijeme).
+            var prijavaInfo = await auth.getInfo();
+            List<PokretResponse> list = await db.SkeniraniPokreti.Where(sp=> sp.KorisnikId == 
+            prijavaInfo.Prijava.KorisnikId).OrderByDescending(sp=> sp.DatumVrijeme).
                 Select(sp => new PokretResponse
                 { Id = sp.Id, Datum = sp.DatumVrijeme.Date.ToString(),
                     Vrijeme = $"{sp.DatumVrijeme.TimeOfDay.Hours}:{sp.DatumVrijeme.TimeOfDay.Minutes}",
@@ -48,7 +55,9 @@ namespace SmartHomeApi.Controllers
         [HttpGet]
         public async Task<PokretiResponse> getPokretiByDate ([FromQuery] DateTime datum )
         {
-            List<PokretResponse> list = await db.SkeniraniPokreti.Where(sp => sp.DatumVrijeme.Date == datum.Date)
+            var prijavaInfo = await auth.getInfo();
+            List<PokretResponse> list = await db.SkeniraniPokreti.Where(
+                sp => sp.KorisnikId == prijavaInfo.Prijava.KorisnikId && sp.DatumVrijeme.Date == datum.Date)
                 .OrderByDescending(sp => sp.DatumVrijeme).
                 Select(sp => new PokretResponse
                 { Id = sp.Id,  Datum = sp.DatumVrijeme.Date.ToString(),
@@ -61,8 +70,10 @@ namespace SmartHomeApi.Controllers
         [HttpGet]
         public async Task<PokretiResponse> getPokretiLast7Days ()
         {
+            var prijavaInfo = await auth.getInfo();
             DateTime datum = DateTime.Now.AddDays(-7).Date;
-            var lista = await db.SkeniraniPokreti.Where(sp => sp.DatumVrijeme.Date >= datum)
+            var lista = await db.SkeniraniPokreti.Where(
+                sp => sp.KorisnikId == prijavaInfo.Prijava.KorisnikId && sp.DatumVrijeme.Date >= datum)
                 .OrderByDescending(sp => sp.DatumVrijeme).Select(
                 sp => new PokretResponse
                 {
@@ -75,8 +86,10 @@ namespace SmartHomeApi.Controllers
         [HttpGet]
         public async Task<PokretiResponse> getPokretiLastMonth()
         {
+            var prijavaInfo = await auth.getInfo();
             DateTime datum = DateTime.Now.AddDays(-30).Date;
-            var lista = await db.SkeniraniPokreti.Where(sp => sp.DatumVrijeme.Date >= datum)
+            var lista = await db.SkeniraniPokreti.Where(
+                sp => sp.KorisnikId == prijavaInfo.Prijava.KorisnikId && sp.DatumVrijeme.Date >= datum)
                 .OrderByDescending(sp => sp.DatumVrijeme).Select(
                 sp => new PokretResponse
                 {
@@ -91,6 +104,8 @@ namespace SmartHomeApi.Controllers
         {
             SkeniraniPokret pokret = await db.SkeniraniPokreti.FindAsync(id);
             if (pokret == null) return false;
+            var prijavaInfo = await auth.getInfo();
+            if (pokret.KorisnikId != prijavaInfo.Prijava.KorisnikId) return false;
             db.SkeniraniPokreti.Remove(pokret);
             db.SaveChanges();
             return true;
